Metrics Bug Fix

+++ s6-spring-boot-advanced-actuator/spring-boot-advanced-actuator-start/src/main/java/com/thinkenterprise/service/RouteService.java
@@ -25,9 +25,9 @@ import org.springframework.stereotype.Service;
 public class RouteService {
 
 	
-	public Long businessFunction() {
+	public Integer businessFunction() {
 
-	return 0L;
+	return 0;
 	}
 	
 }
